+++ joulescope_ui/widgets/waveform/test/test_axis_ticks.py
@@ -19,6 +19,7 @@ Test the axis ticks
 import unittest
 from joulescope_ui.widgets.waveform import axis_ticks as t
 from joulescope_ui import time64
+import numpy as np
 
 
 class TestAxisTicksXOffset(unittest.TestCase):
@@ -67,12 +68,12 @@ class TestAxisXTicks(unittest.TestCase):
     def test_bad(self):
         x0, x1 = 180027262723932739, 180027262785536828
         v = t.x_ticks(x0, x1, 13.857142857142858)
-        print((x0 - v['offset']) / time64.SECOND)
+        #print((x0 - v['offset']) / time64.SECOND)
         z0 = v['offset'] + int(v['major'][0] * time64.SECOND)
         z1 = v['offset'] + int(v['major'][-1] * time64.SECOND)
         self.assertGreaterEqual(z0, x0)
         self.assertLessEqual(z1, x1)
-        print(v)
+        #print(v)
 
 
 class TestTimeFormat(unittest.TestCase):
@@ -92,3 +93,28 @@ class TestTimeFormat(unittest.TestCase):
         self.assertEqual(('1', 'd'), t.time_fmt(day, 3 * day, day))
         self.assertEqual(('2', 'd'), t.time_fmt(2 * day, 3 * day, day))
         self.assertEqual(('1:00:00:01', 'd:hh:mm:ss'), t.time_fmt(day + 1, 3 * day, 1))
+
+
+class TestAxisTicks(unittest.TestCase):
+
+    def test_normal(self):
+        ticks = t.ticks(0, 1.01, 0.2, 10)
+        np.testing.assert_allclose(np.linspace(0.0, 1.0, 6), ticks['major'])
+        self.assertEqual(0.2, ticks['major_interval'])
+        self.assertEqual(0.02, ticks['minor_interval'])
+        self.assertEqual('', ticks['unit_prefix'])
+
+    def test_major_max(self):
+        ticks = t.ticks(0, 1.01, major_max=3)
+        np.testing.assert_allclose(np.linspace(0.0, 1.0, 3), ticks['major'])
+
+    def test_major_max_override(self):
+        ticks = t.ticks(0, 1.01, 0.2, 3)
+        np.testing.assert_allclose(np.linspace(0.0, 1.0, 3), ticks['major'])
+
+    def test_preferred_prefix(self):
+        ticks = t.ticks(0, 1.01, 0.2, prefix_preferred='m')
+        np.testing.assert_allclose(np.linspace(0.0, 1.0, 6), ticks['major'])
+        self.assertEqual(['0', '200', '400', '600', '800', '1000'], ticks['labels'])
+        self.assertEqual(0.2, ticks['major_interval'])
+        self.assertEqual('m', ticks['unit_prefix'])